added decimal logic, need to add unit test that makes sure we get the percentage in our owner fee

+++ src/state.rs
@@ -1,4 +1,4 @@
-use cosmwasm_std::{Addr, Coin};
+use cosmwasm_std::{Addr, Coin, Decimal};
 use cw_storage_plus::{Item, Map};
 use cw_utils::Expiration;
 use schemars::JsonSchema;
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 
 pub const ADMIN: Item<Addr> = Item::new("admin_addr");
 
-pub const HOUSE_FEE: Item<u64> = Item::new("house_fee");
+pub const HOUSE_FEE: Item<Decimal> = Item::new("house_fee");
 
 // Map of players and their ticket allocation
 pub const PLAYERS: Map<&Addr, PlayerInfo> = Map::new("players");
